+++ podman/tests/unit/test_010_version.py
@@ -1,10 +1,13 @@
+"""system unit tests"""
 import unittest
 
-import podman
 from podman import ApiConnection, system
 
 
 class TestSystem(unittest.TestCase):
+    """Test the system calls"""
+
     def test_000_version(self):
+        """test version call"""
         with ApiConnection("ssh://localhost//") as api:
             self.assertRaises(NotImplementedError, system.version, api)
