+++ tox.ini
@@ -1,0 +1,23 @@
+[tox]
+minversion = 3.2.0
+envlist = py36,py38,py39,pylint,coverage
+ignore_basepython_conflict = true
+
+[testenv]
+basepython = python3
+usedevelop = True
+install_command = pip install {opts} {packages}
+deps = -r{toxinidir}/test-requirements.txt
+commands = nosetests
+
+[testenv:venv]
+commands = {posargs}
+
+[testenv:pylint]
+allowlist_externals = pylint
+commands = pylint podman
+
+[testenv:coverage]
+commands =
+    coverage run -m nose
+    coverage report --fail-under=50 --omit=podman/tests/* --omit=.tox/*
